perf(patch-channel): pay-for-use — driver and emitters shake out of non-patch apps
The list driver + patchDriver were hard-wired into web's client.ts
(insert called driveList directly) and the store write paths imported
patch.ts's emitters statically — every app paid ~2.4 KB brotli
(simple-app floor 10.41 -> 12.99), stores ~1 KB, and 5 of 8 size-limit
scenarios failed.

Now: web's driver lives in patch-driver.ts and arms an insert hook slot
lazily from rowProof/patchDriver (module-scope install won't do — the
dist is a flat bundle, so a top-level call is an unshakeable side
effect); signals' emitters install as hooks (patch-hooks.ts) at first
registration, sound because every emission site is pc-guarded and pc
only exists via registration. registerSlotPatchNext moved to patch.ts so
slot-only registrations arm too.

Size gate green on all 8 scenarios: app floors back to ~next (10.51 vs
10.41 — the ~100 B seam), store floors +~490 B of trap/walk seams
(limits ratcheted with notes). Driver engagement unchanged: dbmon tick
1.60, full suite 32/32.
