+++ packages/signals/src/store/index.ts
@@ -30,13 +30,13 @@ export { createProjectionNext as createProjection } from "./next/projection.js";
 // dev-build artifact and is undocumented).
 // Compiler-contract surface (DESIGN-PATCH-CHANNEL.md): what patch-mode
 // compiled output links against. Undocumented as an application API.
-export { registerPatch, registerRowOps, patchableRaw } from "./next/patch.js";
 export {
+  registerPatch,
+  registerRowOps,
   registerSlotPatchNext as registerSlotPatch,
-  storeIsShallow,
-  storeHasFamily,
-  storeHasOptimisticFamily
-} from "./next/store.js";
+  patchableRaw
+} from "./next/patch.js";
+export { storeIsShallow, storeHasFamily, storeHasOptimisticFamily } from "./next/store.js";
 export { createOptimisticStoreNext as createOptimisticStore } from "./next/optimistic.js";
 
 /** Public createStore: plain form `(init, options?)` and derived writable
