+++ packages/signals/src/store/next/reconcile.ts
@@ -19,7 +19,12 @@
  *   members (R11). Kind changes replace wholesale (R10).
  */
 import { isEqual } from "../../core/index.js";
-import { emitPatchLocal, emitRowOps, emitSlotPatch, type RowOps } from "./patch.js";
+import type { RowOps } from "./patch.js";
+// Patch-channel emission rides installed hooks (patch-hooks.ts) — this
+// module must never import patch.js at runtime (patch.js imports
+// emitSetterRowOps from here, and the hooks are what keep the channel
+// tree-shakeable for non-patch apps). All calls are `t.pc`-guarded.
+import { patchHooks } from "./patch-hooks.js";
 import {
   $PROXY,
   $TARGET,
@@ -137,7 +142,7 @@ function applyAdopt(t: StoreNextTarget, incoming: any, keyFn: KeyFn | null, proj
   // visits parents before children, so ancestors emitted already. EAGER
   // only — family targets' visibility moment is their fold commit
   // (drainFolds emits there; emitting here too would double-fire).
-  if (eager && t.pc !== null && t.pc.p !== null) emitPatchLocal(t, incoming, old);
+  if (eager && t.pc !== null && t.pc.p !== null) patchHooks!.emitPatchLocal(t, incoming, old);
   // Shallow adoption: records are slot values — sticky raw-mark the incoming
   // set (R41) and never descend; slot notification is the positional diff.
   if (shallow) markRawIngest(incoming);
@@ -283,7 +288,7 @@ function applyAdopt(t: StoreNextTarget, incoming: any, keyFn: KeyFn | null, proj
         // clear-then-refill and pure appends crashed the driver.
         if (sp !== null && i < dlen && (keyFn === null || keyAligned)) {
           const pvS = prevRows[i];
-          if (pvS !== nvP) emitSlotPatch(t, i, nvP, pvS);
+          if (pvS !== nvP) patchHooks!.emitSlotPatch(t, i, nvP, pvS);
         }
         if (!shallow && i < dlen && nvP !== null && typeof nvP === "object")
           descend(unwrapValue(prevRows[i]), nvP, keyFn, fam, proj);
@@ -428,7 +433,7 @@ const hasOwnP = Object.prototype.hasOwnProperty;
 const identityKey = (r: any) => unwrapValue(r);
 export function emitSetterRowOps(t: StoreNextTarget, prevRows: any[], nextRows: any[]): void {
   const ops = buildIdentityRowOps(prevRows, nextRows);
-  if (ops !== null) emitRowOps(t, nextRows, ops);
+  if (ops !== null) patchHooks!.emitRowOps(t, nextRows, ops);
 }
 
 /** Identity-keyed structural diff, returned rather than emitted: shared by
@@ -453,7 +458,7 @@ function buildAndEmitRowOps(
   structStart: number,
   keyFn: KeyFn | null
 ): void {
-  emitRowOps(t, nextRows, buildRowOps(prevRows, nextRows, structStart, keyFn));
+  patchHooks!.emitRowOps(t, nextRows, buildRowOps(prevRows, nextRows, structStart, keyFn));
 }
 
 function buildRowOps(
