+++ packages/web/src/patch-driver.ts
@@ -1,0 +1,504 @@
+// @ts-nocheck
+// Patch-mode dual driver + list driver (DESIGN-PATCH-CHANNEL.md).
+//
+// PAY-FOR-USE: this module is retained ONLY by compiled patch-mode output —
+// `patchDriver`/`rowProof` imports exist solely in apps built with the
+// compiler's patch mode on. Importing it installs the list driver into the
+// runtime's insert seam (`installListDriver`); classic apps retain nothing
+// but an undefined check. Do not import this module from the always-retained
+// runtime graph.
+import {
+  createOwner,
+  onCleanup,
+  patchableRaw,
+  registerPatch,
+  registerRowOps,
+  registerSlotPatch,
+  runWithOwner,
+  sharedConfig,
+  storeHasOptimisticFamily,
+  storeIsShallow,
+  untrack
+} from "solid-js";
+import { effect } from "./render.js";
+import { installListDriver } from "./client.js";
+
+const PURE_ROW = Symbol.for("solid.pure-row");
+
+// SIDE-EFFECT-FREE ARMING: the dist is a flat bundle, so a module-scope
+// install call would be an unshakeable top-level side effect retaining the
+// whole driver in every app. Instead the compiled entry points arm the
+// insert seam on first call — rowProof runs at template creation, always
+// before the list's insert; both are pure named exports that consumers
+// without patch-mode output never reference and fully shake.
+const arm = () => {
+  installListDriver(driveList);
+};
+
+export function rowProof<T extends Function>(fn: T): T {
+  arm();
+  (fn as any)[PURE_ROW] = true;
+  return fn;
+}
+
+// Patch-mode dual driver (DESIGN-PATCH-CHANNEL.md, PR-C): compiled template
+// scopes whose bindings are pure member reads of one subject hand ONE
+// compiled body `(next, prev, force) => { compares + writes }` here.
+// - Patchable store record: initial force-apply reads the RAW backing (no
+//   proxy traffic, no tracking), then the store's own visibility transitions
+//   dispatch the body through the patch channel (effect-phase timing, lanes,
+//   transition holds — all channel semantics).
+// - Anything else (props, signals-derived objects, accessor records): a
+//   render effect force-applies the same body; reads through the subject
+//   track normally, force short-circuits every compare so `prev` is never
+//   dereferenced. Same semantics, different dispatcher.
+// Row-bind collector, active while the list driver binds a row.
+// - `unbinds`: every patch registration made during the bind (deep rows —
+//   the stamped template's one patchDriver on the row record). The driver
+//   retains them per row so a REMOVED row's registration is severed even
+//   when user code externally retains the record — otherwise the patch
+//   keeps firing against detached DOM for the record's lifetime, where
+//   classic per-row effects die with the row (audit lifecycle hole).
+// - `bodies`: shallow store rows are RAW (no record target to register on),
+//   so compiled bodies whose subject IS the row are collected and
+//   dispatched by the driver from the array's slot-patch channel.
+let rowCollector: { row: any; bodies: any[]; unbinds: (() => void)[] } | null = null;
+
+// Longest-increasing-subsequence over row-ops sources: positions whose rows
+// are already in relative order (they stay put; everything else moves).
+// Standard patience-sort with predecessor links; -1 sources (new rows) are
+// not part of the sequence.
+const lisPositions = (sources: number[]) => {
+  const n = sources.length;
+  const tails: number[] = [];
+  const tailsIdx: number[] = [];
+  const prev = new Array(n).fill(-1);
+  for (let j = 0; j < n; j++) {
+    const v = sources[j];
+    if (v === -1) continue;
+    let lo = 0,
+      hi = tails.length;
+    while (lo < hi) {
+      const mid = (lo + hi) >> 1;
+      if (tails[mid] < v) lo = mid + 1;
+      else hi = mid;
+    }
+    if (lo > 0) prev[j] = tailsIdx[lo - 1];
+    tails[lo] = v;
+    tailsIdx[lo] = j;
+  }
+  const stable = new Set<number>();
+  let k = tailsIdx.length ? tailsIdx[tails.length - 1] : -1;
+  while (k >= 0) {
+    stable.add(k);
+    k = prev[k];
+  }
+  return stable;
+};
+
+// Patch-mode list driver (DESIGN-PATCH-CHANNEL §3b): drives a keyed store
+// array structurally through registerRowOps — create/bind at op-apply, LIS
+// moves, node removal — bypassing mapArray and the second (DOM-side) diff.
+// Called by the runtime's insert when a `<For>` accessor carries `$ll`
+// metadata; returns false to decline (unproven row function, non-store
+// subject, hydration mismatch), in which case insert falls through to the
+// classic mapArray path by simply calling the accessor.
+//
+// Row purity is proven at COMPILE time (§3c): the driver engages only for
+// row functions carrying the compiler's `rowProof` stamp — one compiled
+// template, no reactive or owned work, patches only on the row parameter.
+// Rows therefore need no per-row owners: value updates ride each record's
+// registered patch, structure rides the array's row-ops channel, and a
+// removed row's registrations die with its record. There is no speculative
+// build and no runtime probe; `lateClassic` only serves ENGAGED lists whose
+// subject later leaves the contract (identity swap to a derived array, a
+// shallow<->deep kind switch).
+export const driveList = (parent: Node, listFn: any, marker?: Node, lateClassic?: () => void) => {
+  const meta = listFn.$ll;
+  // Compile-time admission: unstamped row functions never engage.
+  if (meta.row?.[PURE_ROW] !== true) return false;
+  // `keyed={fn}` rows receive ACCESSORS (the classic contract) — the driver
+  // binds rows with raw records, so engaging would hand user code the wrong
+  // shape. Decline until the accessor-row binding + compiler grammar for
+  // `param().member` bodies lands (identity-ruling follow-up); these rows
+  // cannot currently stamp anyway, this is a defensive contract pin.
+  if (typeof meta.keyed === "function") return false;
+  // The decision read is id-ISOLATED: evaluating `each` can mint compiler
+  // memos lazily inside the prop getter (wrapConditionals), and minting them
+  // on the ambient chain here would consume a child id the classic path
+  // expects to consume later — shifting every subsequent hydration key on
+  // decline. A throwaway explicit-id owner absorbs (and disposal discards)
+  // anything the read creates.
+  const evalOwner = createOwner({ id: "&each" });
+  let subject: any = runWithOwner(evalOwner, () => untrack(meta.each));
+  (evalOwner as any).dispose();
+  let raw = subject != null ? patchableRaw(subject) : undefined;
+  if (raw === undefined || !Array.isArray(raw)) return false;
+  // OPTIMISTIC families engage (family increment 2): their structural
+  // writes emit lane-timed row ops from the override channel, and reverts
+  // emit an identity RESYNC (ops === null). The committed backing lags the
+  // visible state while optimism is in flight, so the initial bind reads
+  // the OPTIMISTIC VIEW through the proxy — classic mapArray reads the same
+  // view, and the equivalence matrix holds across writes/reverts/landings.
+  const optimistic = storeHasOptimisticFamily(subject);
+  if (optimistic) raw = untrack(() => Array.from(subject as any));
+
+  // Hydration precheck (claim + register only — §5): rows are the region's
+  // server-rendered elements, claimed positionally through each element's
+  // own `_hk` key. V1 supports the whole-parent region (no marker) and
+  // requires an exact row count and a clean key on every row; anything else
+  // declines to classic hydration. Keys end in the row scope's FIRST child
+  // id ("0" — pure rows consume no ids before the root claim), so the row
+  // owner's id is the key minus that suffix.
+  const hydrating = !!sharedConfig.hydrating;
+  // Empty-initial lists have nothing to claim — classic hydration owns them.
+  if (hydrating && raw.length === 0) return false;
+  let domRows: Element[] | undefined;
+  let rowIds: string[] | undefined;
+  if (hydrating) {
+    if (marker !== undefined) return false;
+    domRows = Array.from((parent as Element).children);
+    if (domRows.length !== raw.length) return false;
+    rowIds = new Array(raw.length);
+    for (let i = 0; i < domRows.length; i++) {
+      const key = domRows[i].getAttribute("_hk");
+      if (key === null || !key.endsWith("0") || key.length < 2) return false;
+      rowIds[i] = key.slice(0, -1);
+    }
+  }
+
+  const rowFn = meta.row;
+  const endAnchor = marker ?? null;
+
+  // Shallow store lists: rows are RAW, so compiled bodies are COLLECTED at
+  // bind (patchDriver's rowCollector branch) and dispatched from the array's
+  // slot-patch channel; `lastBodies` carries each bind's collection to its
+  // bookkeeping site.
+  const shallow = storeIsShallow(subject);
+  let lastBodies: any[] | null = null;
+  let lastUnbinds: (() => void)[] | null = null;
+  const collectBind = (abs: number, build: () => Node): Node => {
+    const prevC = rowCollector;
+    rowCollector = { row: shallow ? subject[abs] : undefined, bodies: [], unbinds: [] };
+    try {
+      return build();
+    } finally {
+      lastBodies = rowCollector.bodies;
+      lastUnbinds = rowCollector.unbinds;
+      rowCollector = prevC;
+    }
+  };
+
+  // Engaged. The list owner consumes exactly one child id, mirroring the
+  // owner mapArray would have created — subsequent siblings' hydration ids
+  // stay aligned on both the engage and (pre-owner) decline paths.
+  const listOwner = createOwner();
+  let declined = false;
+  const bindRow = (abs: number, claimId?: string): Node => {
+    if ("_DX_DEV_") {
+      // Ownership assertion: a stamped row must attach NOTHING to the list
+      // owner — the compiler proved the template, but handler/attribute
+      // VALUE expressions are arbitrary user code, and owned work created
+      // there (a handler factory calling onCleanup/createEffect) would
+      // outlive the row. Snapshot the owner's slots around the real build.
+      const o = listOwner as any;
+      const prevChild = o._firstChild;
+      const prevDisposal = o._disposal;
+      const node = collectBind(abs, () =>
+        runWithOwner(listOwner, () =>
+          claimId !== undefined
+            ? (runWithOwner(createOwner({ id: claimId }) as any, () =>
+                untrack(() => rowFn(subject[abs]))
+              ) as Node)
+            : (untrack(() => rowFn(subject[abs])) as Node)
+        )
+      ) as Node;
+      if (o._firstChild !== prevChild || o._disposal !== prevDisposal) {
+        console.warn(
+          "A patch-mode list row created reactive computations or cleanups " +
+            "during build (likely a handler/attribute value expression calling " +
+            "createEffect/onCleanup). This work attaches to the LIST, not the " +
+            "row, and will not dispose when the row is removed. Move owned " +
+            "work into effects/refs (which opt the row out of patch mode)."
+        );
+      }
+      return node;
+    }
+    return collectBind(abs, () =>
+      runWithOwner(listOwner, () =>
+        claimId !== undefined
+          ? (runWithOwner(createOwner({ id: claimId }) as any, () =>
+              untrack(() => rowFn(subject[abs]))
+            ) as Node)
+          : (untrack(() => rowFn(subject[abs])) as Node)
+      )
+    ) as Node;
+  };
+
+  let entries: Node[] = new Array(raw.length);
+  let rowBodies: any[][] | null = shallow ? new Array(raw.length) : null;
+  // Per-row patch unbind handles (deep rows register on their record): run
+  // on row removal, contract-leave, and list disposal, so a record the app
+  // retains beyond the row cannot keep patching detached DOM.
+  let rowUnbinds: (() => void)[][] = new Array(raw.length);
+  const runUnbinds = (list: (() => void)[] | undefined) => {
+    if (list !== undefined) for (let u = 0; u < list.length; u++) list[u]();
+  };
+  const unbindAllRows = () => {
+    for (let j = 0; j < rowUnbinds.length; j++) runUnbinds(rowUnbinds[j]);
+    rowUnbinds = [];
+  };
+  let prevRaws: any[] = raw.slice();
+  if (hydrating) {
+    // Claim pass: each bind claims its server row through the row-scoped id
+    // (getNextElement resolves the `_hk` registry entry); patchDriver skips
+    // the initial apply.
+    for (let i = 0; i < raw.length; i++) {
+      entries[i] = bindRow(i, rowIds![i]);
+      if (rowBodies !== null) rowBodies[i] = lastBodies!;
+      rowUnbinds[i] = lastUnbinds!;
+    }
+  } else {
+    for (let i = 0; i < raw.length; i++) {
+      const node = bindRow(i);
+      entries[i] = node;
+      if (rowBodies !== null) rowBodies[i] = lastBodies!;
+      rowUnbinds[i] = lastUnbinds!;
+      parent.insertBefore(node, endAnchor);
+    }
+  }
+
+  // Synthetic full-window ops by ROW IDENTITY against the retained raws:
+  // used by identity swaps (`s.rows = newArr`) and by the optimistic revert
+  // RESYNC (ops === null — overrides are gone, the live view is truth, and
+  // retention must be rebuilt by identity). RAW identity on both sides:
+  // draft-authored permutations produce arrays of row PROXIES and deep
+  // ingest stores them verbatim — matching without unwrapping rebuilds
+  // every row (JFB keyed-reorder identity gate).
+  const identityOps = (nextArr: any[]): { prefix: number; sources: number[] } => {
+    const keyOf = (r: any) => {
+      const w = r != null ? patchableRaw(r) : undefined;
+      return w !== undefined ? w : r;
+    };
+    const oldIndex = new Map<any, number>();
+    for (let j = 0; j < prevRaws.length; j++) {
+      const k = keyOf(prevRaws[j]);
+      if (!oldIndex.has(k)) oldIndex.set(k, j);
+    }
+    const sources = new Array(nextArr.length);
+    for (let k = 0; k < nextArr.length; k++) sources[k] = oldIndex.get(keyOf(nextArr[k])) ?? -1;
+    return { prefix: 0, sources };
+  };
+  const applyOps = (next: any[], ops: { prefix: number; sources: number[] } | null) => {
+    if (declined) return;
+    if (ops === null) ops = identityOps(next);
+    const { prefix, sources } = ops;
+    const retained = new Set<number>();
+    for (let j = 0; j < sources.length; j++) if (sources[j] >= 0) retained.add(sources[j]);
+    for (let j = prefix; j < entries.length; j++) {
+      if (!retained.has(j)) {
+        (entries[j] as ChildNode).remove();
+        runUnbinds(rowUnbinds[j]);
+      }
+    }
+    const newEntries: Node[] = new Array(prefix + sources.length);
+    const newBodies: any[][] | null =
+      rowBodies !== null ? new Array(prefix + sources.length) : null;
+    const newUnbinds: (() => void)[][] = new Array(prefix + sources.length);
+    for (let i = 0; i < prefix; i++) {
+      newEntries[i] = entries[i];
+      if (newBodies !== null) newBodies[i] = rowBodies![i];
+      newUnbinds[i] = rowUnbinds[i];
+    }
+    const stable = lisPositions(sources);
+    let anchor: Node | null = endAnchor;
+    for (let j = sources.length - 1; j >= 0; j--) {
+      const abs = prefix + j;
+      const src = sources[j];
+      let node: Node;
+      if (src === -1) {
+        node = bindRow(abs);
+        if (newBodies !== null) newBodies[abs] = lastBodies!;
+        newUnbinds[abs] = lastUnbinds!;
+        parent.insertBefore(node, anchor);
+      } else if (refRebuild && next[abs] !== prevRaws[src]) {
+        // Reference-keyed shallow lists: a retained (key-matched, moved)
+        // slot whose record was replaced REBUILDS — reference semantics.
+        runUnbinds(rowUnbinds[src]);
+        (entries[src] as ChildNode).remove();
+        node = bindRow(abs);
+        if (newBodies !== null) newBodies[abs] = lastBodies!;
+        newUnbinds[abs] = lastUnbinds!;
+        parent.insertBefore(node, anchor);
+      } else {
+        node = entries[src];
+        if (newBodies !== null) newBodies[abs] = rowBodies![src];
+        newUnbinds[abs] = rowUnbinds[src];
+        if (!stable.has(j)) parent.insertBefore(node, anchor);
+      }
+      newEntries[abs] = node;
+      anchor = node;
+    }
+    entries = newEntries;
+    if (newBodies !== null) rowBodies = newBodies;
+    rowUnbinds = newUnbinds;
+    prevRaws = next.slice();
+  };
+
+  let unbindOps = runWithOwner(listOwner, () => registerRowOps(subject, applyOps)) as () => void;
+
+  // IDENTITY SEMANTICS RULING: the driver implements whatever identity the
+  // VIEW declared, never the reconcile key's (that would make patch mode a
+  // semantic change, not an optimization — the compiler is default-on).
+  //   - deep lists: adoption preserves proxy identity per key, so key ops
+  //     and reference semantics coincide by construction — nothing to do.
+  //   - shallow + reference-keyed (`keyed` absent/true): the records ARE the
+  //     identity. A key-aligned slot whose record was REPLACED must rebuild
+  //     its row, exactly as classic mapArray does.
+  //   - shallow + `keyed={fn}`: replacement under a matching key is a value
+  //     tick — patch the row in place (the declared semantics).
+  const refRebuild = shallow && typeof meta.keyed !== "function";
+  const rebuildSlot = (i: number): void => {
+    runUnbinds(rowUnbinds[i]);
+    const old = entries[i] as ChildNode;
+    const node = bindRow(i);
+    rowBodies![i] = lastBodies!;
+    rowUnbinds[i] = lastUnbinds!;
+    parent.insertBefore(node, old);
+    old.remove();
+    entries[i] = node;
+  };
+  // Shallow value channel: a key-aligned slot replaced by reference. Under
+  // declared-key semantics this is a value tick — run the row's collected
+  // bodies against (next, prev) and adopt the new raw as that slot's
+  // identity. Under reference semantics it is a REPLACE — rebuild the row.
+  // Structure never lands here (the walk emits misaligned slots as row ops
+  // only).
+  const applySlot = (i: number, next: any, prev: any) => {
+    if (declined) return;
+    if (refRebuild) {
+      rebuildSlot(i);
+      prevRaws[i] = next;
+      return;
+    }
+    const bodies = rowBodies![i];
+    if (bodies !== undefined) {
+      for (let b = 0; b < bodies.length; b++) bodies[b](next, prev, false);
+    }
+    prevRaws[i] = next;
+  };
+  let unbindSlots = shallow
+    ? (runWithOwner(listOwner, () => registerSlotPatch(subject, applySlot)) as () => void)
+    : null;
+
+  // Identity swaps (`s.rows = newArr` without reconcile) keep mapArray's
+  // keyed semantics: rows matched by RAW IDENTITY retain their DOM; the rest
+  // bind/remove through the same LIS apply, as a synthetic full-window op.
+  // Created under the list owner: every tracked `each` read can mint getter
+  // memos, and the list owner's id counter is private (id-chain neutral).
+  runWithOwner(listOwner, () =>
+    effect(
+      () => meta.each(),
+      (value: any) => {
+        if (declined || value === subject) return;
+        const nextRaw = value != null ? patchableRaw(value) : undefined;
+        unbindOps();
+        unbindSlots?.();
+        // A swap that changes the store KIND (shallow <-> deep) leaves this
+        // engagement's channel wiring invalid — treat it like leaving the
+        // contract and hand off to classic.
+        if (nextRaw !== undefined && Array.isArray(nextRaw) && storeIsShallow(value) !== shallow) {
+          for (let j = 0; j < entries.length; j++) (entries[j] as ChildNode).remove();
+          entries = [];
+          prevRaws = [];
+          unbindAllRows();
+          subject = value;
+          declined = true;
+          (listOwner as any).dispose();
+          lateClassic?.();
+          return;
+        }
+        if (nextRaw === undefined || !Array.isArray(nextRaw)) {
+          // Subject left the driver's contract (e.g. `each` switched from
+          // the store array to a DERIVED array — a filtered view). Clear the
+          // region and hand the list to the classic path, which renders the
+          // current subject and owns it from here on.
+          for (let j = 0; j < entries.length; j++) (entries[j] as ChildNode).remove();
+          entries = [];
+          prevRaws = [];
+          unbindAllRows();
+          subject = value;
+          declined = true;
+          (listOwner as any).dispose();
+          lateClassic?.();
+          return;
+        }
+        const swapOps = identityOps(nextRaw);
+        subject = value;
+        applyOps(nextRaw, swapOps);
+        unbindOps = runWithOwner(listOwner, () => registerRowOps(subject, applyOps)) as () => void;
+        if (shallow)
+          unbindSlots = runWithOwner(listOwner, () =>
+            registerSlotPatch(subject, applySlot)
+          ) as () => void;
+      }
+    )
+  );
+  onCleanup(() => {
+    unbindOps();
+    unbindSlots?.();
+    // Sever every row's patch registration, not just the channels: the
+    // channel skips disposed-owner entries but never removes them, so a
+    // record the app retains past the list would otherwise carry dead
+    // entries for its lifetime.
+    unbindAllRows();
+    (listOwner as any).dispose();
+  });
+  return true;
+};
+
+// Patch-mode dual driver: compiled template scopes whose bindings are pure
+// member reads of ONE subject hand a single compiled body
+// `(next, prev, force) => { compares + writes }` here.
+// - Patchable record (core provides the seams): the initial force-apply
+//   reads the raw backing, then the core's own visibility transitions
+//   dispatch the body through its patch channel. Under hydration the
+//   registration alone arms the record — server HTML already carries
+//   current values, so the initial apply is skipped.
+// - Anything else (props, derived objects, unaware cores): a dual-phase
+//   effect runs the same body — the compute pass calls it with
+//   next === prev so every compare fails and it becomes a pure tracked
+//   read; the commit pass force-applies, keeping DOM writes in the effect
+//   phase where transitions and batching expect them.
+export const patchDriver = (subject, body) => {
+  arm();
+  const raw = patchableRaw(subject);
+  if (raw !== undefined) {
+    // Hydration is claim + register ONLY (DESIGN-PATCH-CHANNEL §5): the
+    // server HTML already carries current values, so the initial force-apply
+    // is skipped — no writes, no graph edges. The registration alone arms
+    // the record for post-hydration transitions.
+    if (!sharedConfig.hydrating) body(raw, undefined, true);
+    const unbind = registerPatch(subject, body);
+    if (rowCollector !== null) rowCollector.unbinds.push(unbind);
+  } else if (rowCollector !== null && subject === rowCollector.row) {
+    rowCollector.bodies.push(body);
+    if (!sharedConfig.hydrating) body(subject, undefined, true);
+  } else {
+    // Effect fallback with correct WRITE TIMING: the compute pass calls the
+    // body with next === prev, so every compare fails and it becomes a pure
+    // TRACKED READ of each binding expression (eligible expressions are pure
+    // member chains — double evaluation is free of side effects); the commit
+    // pass force-applies, putting DOM writes in the effect phase where
+    // transitions and batching expect them — same split as classic compiled
+    // effects, same single compiled body.
+    effect(
+      () => body(subject, subject, false),
+      // untrack: the commit pass re-evaluates binding expressions by design
+      // (force short-circuits compares, not reads) — without it, dev-mode
+      // strict-read flags every re-read as an untracked effect-callback read
+      // (false positive: the compute pass tracked the same expressions).
+      () => untrack(() => body(subject, undefined, true))
+    );
+  }
+};
